pyproject.toml
[project]
name = "schemas"
version = "0.1.0"
description = "Schema gen tool"
authors = [
    {name = "Frank1o3",email = "jahdy1o3@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "jsonschema (>=4.26.0,<5.0.0)"
]
packages = [{include = "tools"}]

[dependency-groups]
dev = [
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
